Refactor mm import map
