+++ src/mm.rs
@@ -1,14 +1,5 @@
-pub mod meta_page;
-
-mod page_allocator;
-pub use page_allocator::{util::*, Page, PageAllocator, GFP, PAGE_ALLOC};
-pub mod virtual_allocator;
-
-pub mod cache_allocator;
+pub mod alloc;
 pub mod constant;
-pub mod global_allocator;
-pub mod memory_allocator;
+pub mod init;
+pub mod page;
 pub mod util;
-pub mod x86;
-
-pub mod kmalloc;
