+++ src/mm/alloc/cache.rs
@@ -1,0 +1,11 @@
+mod cache_allocator;
+mod cache_manager;
+mod meta_cache;
+mod no_alloc_list;
+mod size_cache;
+mod traits;
+
+// pub use memory_allocator::
+pub use cache_allocator::{CacheAllocator, CacheAllocatorStat};
+pub use cache_manager::CM;
+pub use traits::{CacheInit, CacheStat, CacheTrait};
