+++ src/mm/alloc/page.rs
@@ -1,0 +1,5 @@
+mod buddy_allocator;
+mod free_list;
+mod page_allocator;
+
+pub use page_allocator::{PageAlloc, PAGE_ALLOC};
