+++ src/mm/alloc/phys.rs
@@ -1,0 +1,8 @@
+mod atomic;
+mod global;
+mod memory_allocator;
+mod normal;
+
+pub use atomic::MemAtomic;
+pub use memory_allocator::PMemAlloc;
+pub use normal::MemNormal;
