--- src/mm/virtual_allocator.rs
@@ -1,1 +1,0 @@
-pub mod vmalloc;
