--- src/mm/x86.rs
@@ -1,3 +1,0 @@
-pub mod init;
-pub mod util;
-pub mod x86_page;
